Renames rule to `prefer-event-key`
