+++ index.js
@@ -40,6 +40,7 @@ module.exports = {
 				'unicorn/number-literal-case': 'error',
 				'unicorn/prefer-add-event-listener': 'error',
 				'unicorn/prefer-key-over-key-code': 'error',
+				'unicorn/prefer-event-key': 'error',
 				'unicorn/prefer-exponentiation-operator': 'error',
 				'unicorn/prefer-includes': 'error',
 				'unicorn/prefer-node-append': 'error',
